safari-600.7-branch/LayoutTests/fast/dom/DOMURL/set-href-attribute-whitespace.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>

description('Test setting the href attribute of a URL object to a URL with leading and trailing whitespace.');

var a = new URL("about:blank");

debug("Set href that starts with a space");
a.href = " https://www.mydomain.com/path/testurl.html?key=value";
shouldBe("a.hostname", "'www.mydomain.com'");

debug("Set href that starts with a newline");
a.href = "\nhttps://www.mydomain.com/path/testurl.html?key=value";
shouldBe("a.hostname", "'www.mydomain.com'");

debug("Set href that starts with a tab");
a.href = "\thttps://www.mydomain.com/path/testurl.html?key=value";
shouldBe("a.hostname", "'www.mydomain.com'");

debug("Set href that starts with a carriage return");
a.href = "\rhttps://www.mydomain.com/path/testurl.html?key=value";
shouldBe("a.hostname", "'www.mydomain.com'");

debug("Set href that starts with a combination of newlines, spaces and tabs");
a.href = "\n \t\r \nhttps://www.mydomain.com/path/testurl.html?key=value";
shouldBe("a.hostname", "'www.mydomain.com'");

debug("Set href that ends with a space");
a.href = "https://www.mydomain.com/path/testurl.html?key=value ";
shouldBe("a.hostname", "'www.mydomain.com'");

debug("Set href that ends with a newline");
a.href = "https://www.mydomain.com/path/testurl.html?key=value\n";
shouldBe("a.hostname", "'www.mydomain.com'");

debug("Set href that ends with a tab");
a.href = "https://www.mydomain.com/path/testurl.html?key=value\t";
shouldBe("a.hostname", "'www.mydomain.com'");

debug("Set href that ends with a carriage return");
a.href = "https://www.mydomain.com/path/testurl.html?key=value\r";
shouldBe("a.hostname", "'www.mydomain.com'");

debug("Set href that ends with a combination of newlines, spaces and tabs");
a.href = "https://www.mydomain.com/path/testurl.html?key=value\n \t\r \n";
shouldBe("a.hostname", "'www.mydomain.com'");

debug("Set href that starts and ends with a combination of newlines, spaces and tabs");
a.href = "\n \t\r \nhttps://www.mydomain.com/path/testurl.html?key=value\n \t\r \n";
shouldBe("a.hostname", "'www.mydomain.com'");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
